add token on all controllers

+++ src/main/java/com/example/backend/controllers/CategoryController.java
@@ -1,8 +1,11 @@
 package com.example.backend.controllers;
 
+import com.example.backend.Config.TokenService;
 import com.example.backend.model.Category;
 import com.example.backend.services.Category.CategoryService;
 import org.bson.types.ObjectId;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -13,19 +16,28 @@ import java.util.List;
 public class CategoryController {
 
     private CategoryService categoryService;
+    private TokenService tokenService;
 
-    public CategoryController(CategoryService categoryService) {
+    public CategoryController(CategoryService categoryService, TokenService tokenService) {
         this.categoryService = categoryService;
+        this.tokenService = tokenService;
     }
 
     @GetMapping
-    public List<Category> getAllCategories(){
-        return categoryService.getAllCategories();
+    public ResponseEntity<List<Category>> getAllCategories(@RequestHeader(name="Authorization")String token){
+        if(tokenService.verifyToken(token)){
+            return ResponseEntity.ok().body(categoryService.getAllCategories());
+        }
+        return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
     }
 
     @GetMapping("/{id}")
-    public Category getOneCategory(@PathVariable String id){
-        return categoryService.getOneCategory(new ObjectId(String.valueOf(id)));
+    public ResponseEntity<Category> getOneCategory(@RequestHeader(name="Authorization")String token,@PathVariable String id){
+        if(tokenService.verifyToken(token)){
+            return ResponseEntity.ok().body(categoryService.getOneCategory(new ObjectId(String.valueOf(id))));
+        }
+        return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+
     }
 
     //pas nécessaire
@@ -47,18 +59,25 @@ public class CategoryController {
     }*/
 
     @PostMapping
-    public Category addCategory(@RequestBody Category category){
+    public ResponseEntity<Category> addCategory(@RequestHeader(name="Authorization")String token,@RequestBody Category category){
         Category c=null;
-        try{
-            c=categoryService.saveCategory(category);
-        }catch(Exception e){
-            e.printStackTrace();
+        if(tokenService.verifyTokenAndAdmin(token)) {
+            try {
+                c = categoryService.saveCategory(category);
+                return ResponseEntity.ok().body(c);
+            } catch (Exception e) {
+                e.printStackTrace();
+            }
         }
-        return c;
+        return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
     }
 
     @GetMapping("/delete/{id}")
-    public void deleteCategory(@PathVariable String id){
-        categoryService.deleteCategory(new ObjectId(String.valueOf(id)));
+    public ResponseEntity deleteCategory(@RequestHeader(name="Authorization")String token,@PathVariable String id){
+        if(tokenService.verifyTokenAndAdmin(token)) {
+            categoryService.deleteCategory(new ObjectId(String.valueOf(id)));
+            return ResponseEntity.ok().build();
+        }
+        return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
     }
 }
