+++ src/main/java/com/example/backend/controllers/ProductController.java
@@ -1,6 +1,9 @@
 package com.example.backend.controllers;
+import com.example.backend.Config.TokenService;
 import com.example.backend.Enums.State;
 import org.bson.types.ObjectId;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import com.example.backend.model.Product;
@@ -12,30 +15,47 @@ import org.springframework.web.bind.annotation.*;
 @RequestMapping("/products")
 public class ProductController {
     private ProductService productService;
+    private TokenService tokenService;
 
-    public ProductController(ProductService productService) {
+    public ProductController(ProductService productService,TokenService tokenService) {
         this.productService = productService;
+        this.tokenService = tokenService;
     }
 
     @GetMapping
-    public Iterable<Product> getAllProducts(){
-        return productService.findAllProduct();
+    public ResponseEntity<Iterable<Product>> getAllProducts(@RequestHeader(name="Authorization")String token){
+        if(tokenService.verifyToken(token)){
+            return ResponseEntity.ok().body(productService.findAllProduct());
+        }
+        return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
     }
 
     @GetMapping("/filter/{idSubCategory}")
-    public Iterable<Product> getProductsBySubCategory(@PathVariable String idSubCategory){
-        System.out.println("here is " + idSubCategory );
-        return productService.findAllProductsByIdSubCategory(new ObjectId(String.valueOf(idSubCategory)));
+    public ResponseEntity<Iterable<Product>> getProductsBySubCategory(@RequestHeader(name="Authorization")String token,@PathVariable String idSubCategory){
+        if(tokenService.verifyToken(token)){
+            return ResponseEntity.ok().body(productService.findAllProductsByIdSubCategory(new ObjectId(String.valueOf(idSubCategory))));
+        }
+        return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+
     }
 
     @GetMapping("/{idProduct}")
-    public Product getProduct(@PathVariable String idProduct){
-        return productService.findOneById(new ObjectId(String.valueOf(idProduct)));
+    public ResponseEntity<Product> getProduct(@RequestHeader(name="Authorization")String token,@PathVariable String idProduct){
+        if(tokenService.verifyToken(token)) {
+            return ResponseEntity.ok().body(productService.findOneById(new ObjectId(String.valueOf(idProduct))));
+
+        }
+        return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
     }
 
     @GetMapping("/delete/{idProduct}")
-    public void deleteProduct(@PathVariable("idProduct") String idProduct){
-        productService.deleteProduct(new ObjectId(String.valueOf(idProduct)));
+    public ResponseEntity deleteProduct(@RequestHeader(name="Authorization")String token,@PathVariable("idProduct") String idProduct){
+        if(tokenService.verifyToken(token)){
+            productService.deleteProduct(new ObjectId(String.valueOf(idProduct)));
+            return ResponseEntity.ok().build();
+        }
+        return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+
     }
     /*@PostMapping
     public ResponseEntity<Void> createAdvertisement(@RequestBody Product product) {
@@ -50,41 +70,66 @@ public class ProductController {
     }*/
 
     @PostMapping
-    public Product addProduct(@RequestBody Product product) {
+    public ResponseEntity<Product> addProduct(@RequestHeader(name="Authorization")String token, @RequestBody Product product) {
         Product pro= null;
-        try{
-            product.setState(State.Debut);
-            product.setStateName(product.getState().getName());
-            product.setStatusName(product.getStatus().getName());
-            pro=productService.saveAdvertisement(product);
-        }catch (Exception e){
-            e.printStackTrace();
+        if(tokenService.verifyToken(token)){
+            try{
+                product.setState(State.Debut);
+                product.setStateName(product.getState().getName());
+                product.setStatusName(product.getStatus().getName());
+                pro=productService.saveAdvertisement(product);
+                return ResponseEntity.ok().body(pro);
+            }catch (Exception e){
+                e.printStackTrace();
+            }
         }
-        return pro;
+        return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
     }
 
     @PutMapping ("/{id}")
-    public void updateProduct(@PathVariable ("id") String idProduct,@RequestBody Product product) {
-        productService.updateProduct(new ObjectId(String.valueOf(idProduct)), product);
+    public ResponseEntity<Product> updateProduct(@RequestHeader(name="Authorization")String token,@PathVariable ("id") String idProduct,@RequestBody Product product) {
+        if(tokenService.verifyToken(token)){
+            Product p=productService.updateProduct(new ObjectId(String.valueOf(idProduct)), product);
+            return ResponseEntity.ok().body(p);
+        }
+        return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
     }
 
     @GetMapping("/valid/{id}")
-    public void acceptProduct(@PathVariable String idProduct){
-        productService.acceptProduct(new ObjectId(String.valueOf(idProduct)));
+    public ResponseEntity acceptProduct(@RequestHeader(name="Authorization")String token,@PathVariable String idProduct){
+        if(tokenService.verifyTokenAndAdmin(token)){
+            productService.acceptProduct(new ObjectId(String.valueOf(idProduct)));
+            return ResponseEntity.ok().build();
+        }
+        return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+
     }
 
     @GetMapping("/sell/{id}")
-    public void sellProduct(@PathVariable String idProduct){
-        productService.sellProduct(new ObjectId(String.valueOf(idProduct)));
+    public ResponseEntity sellProduct(@RequestHeader(name="Authorization")String token,@PathVariable String idProduct){
+        if(tokenService.verifyToken(token)){
+            productService.sellProduct(new ObjectId(String.valueOf(idProduct)));
+            return ResponseEntity.ok().build();
+        }
+        return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+
     }
 
     @GetMapping("refuse/{id}")
-    public void refuseProduct(@PathVariable String idProduct){
-        productService.refuseProduct(new ObjectId(String.valueOf(idProduct)));
+    public ResponseEntity refuseProduct(@RequestHeader(name="Authorization")String token,@PathVariable String idProduct){
+        if(tokenService.verifyTokenAndAdmin(token)){
+            productService.refuseProduct(new ObjectId(String.valueOf(idProduct)));
+            return ResponseEntity.ok().build();
+        }
+        return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
     }
 
     @GetMapping("setToDelete/{id}")
-    public void setToDeleteProducts(@PathVariable String idProduct){
-        productService.setToDeleteProducts(new ObjectId(String.valueOf(idProduct)));
+    public ResponseEntity setToDeleteProducts(@RequestHeader(name="Authorization")String token,@PathVariable String idProduct){
+        if(tokenService.verifyTokenAndAdmin(token)){
+            productService.setToDeleteProducts(new ObjectId(String.valueOf(idProduct)));
+            return ResponseEntity.ok().build();
+        }
+        return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
     }
 }
